Created components folder and re-organize style files
