+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
-import { hex } from "@styles/hex";
-import { Button } from "@styles/buttons";
+import { hex } from "@utils/hex";
+import DefaultButton from "@components/Button";
 import type { NextPage } from "next";
 import { signIn, signOut, useSession } from "next-auth/react";
 import Head from "next/head";
@@ -16,9 +16,9 @@ const Home: NextPage = () => {
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
-      <Button theme={{ color: { primary: hex("#f00000") } }}>
+      <DefaultButton theme={{ color: { primary: hex("#f00000") } }}>
         Demo Button
-      </Button>
+      </DefaultButton>
 
       {status === "authenticated" && session.user ? (
         <div>
@@ -39,7 +39,7 @@ const Home: NextPage = () => {
       ) : (
         <>
           <h1>Bruh, sign in??</h1>
-          <Button onClick={() => signIn()}>Sign in</Button>
+          <DefaultButton onClick={() => signIn()}>Sign in</DefaultButton>
         </>
       )}
     </>
